MB-31266: Pass a isDropped(key) callback to ephemeral purger
The ephemeral purger will perform the same function as the
compactor in persistent buckets for the processing of dropped
collections.

All keys are checked to see if they are owned by a dropped
collection by a callback to the KVBucket::collectionsEraseKey
wired through to the purger.

Tests are now executed against persistent and ephmeral buckets
in all eviction modes.

Reviewed-on: http://review.couchbase.org/103785
Tested-by: Chris Farman <christopher.farman@couchbase.com>

+++ engines/ep/src/collections/eraser_context.h
@@ -62,7 +62,7 @@ public:
             std::function<void(cb::const_byte_buffer)> saveManifestCb);
 
 private:
-    int collectionsErased = 0;
+    size_t collectionsErased = 0;
 
     /// The manifest which collection erasing can compare keys.
     Manifest& manifest;
